style(store): rename files with non-standard naming convention
