+++ packages/store/src/public_api.ts
@@ -24,4 +24,4 @@ export { Selector } from './decorators/selector';
 export { getActionTypeFromInstance, actionMatcher } from './utils/utils';
 export { createSelector } from './utils/selector-utils';
 export { NgxsExecutionStrategy } from './execution/symbols';
-export { NoopNgxsExecutionStrategy } from './execution/noopNgxsExecutionStrategy';
+export { NoopNgxsExecutionStrategy } from './execution/noop-ngxs-execution-strategy';
